# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.RegisterRequest:
    properties:
      business_name:
        maxLength: 255
        minLength: 2
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
      rfc:
        maxLength: 13
        minLength: 12
        type: string
    required:
    - business_name
    - email
    - password
    - phone
    - rfc
    type: object
  entities.RegisterResponse:
    properties:
      business_name:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      rfc:
        type: string
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@visionprice.app
    name: VisionPrice Support
  description: API backend para el módulo de Proveedores de VisionPrice
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VisionPrice Proveedores API
  version: "1.0"
paths:
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Crea una nueva cuenta de proveedor con nombre comercial, RFC, correo,
        teléfono y contraseña
      parameters:
      - description: Datos de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Registrar un nuevo proveedor
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Ingrese el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"